♻️(frontend) extract recording api into a dedicated folder
Extract code elements related to recording into a dedicated folder.
