+++ src/frontend/src/features/recording/types.ts
@@ -1,0 +1,4 @@
+export enum RecordingMode {
+  Transcript = 'transcript',
+  ScreenRecording = 'screen_recording',
+}
