+++ src/frontend/src/features/rooms/livekit/components/Transcript.tsx
@@ -7,8 +7,12 @@ import { useRoomContext } from '@livekit/components-react'
 import {
   RecordingMode,
   useStartRecording,
-} from '@/features/rooms/api/startRecording'
-import { useStopRecording } from '@/features/rooms/api/stopRecording'
+  useStopRecording,
+  useIsScreenRecordingStarted,
+  useIsTranscriptStarted,
+  useIsRecordingTransitioning,
+  useHasRecordingAccess,
+} from '@/features/recording'
 import { useEffect, useMemo, useState } from 'react'
 import { RoomEvent } from 'livekit-client'
 import { useTranslation } from 'react-i18next'
@@ -19,12 +23,6 @@ import {
   BETA_USERS_FORM_URL,
   CRISP_HELP_ARTICLE_TRANSCRIPT,
 } from '@/utils/constants'
-import {
-  useIsScreenRecordingStarted,
-  useIsTranscriptStarted,
-  useIsRecordingTransitioning,
-  useHasRecordingAccess,
-} from '@/features/recording'
 
 export const Transcript = () => {
   const [isLoading, setIsLoading] = useState(false)
